refactor: Extract and use payslip module
So we can later reuse it to fetch childminder's payslips.
Also added tests.
